mpc-server/requirements.txt
flask==3.0.0
gunicorn==21.2.0
azure-identity==1.15.0
azure-monitor-query==1.3.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.8.0
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0
pydantic==2.5.0
requests==2.31.0
python-jose[cryptography]==3.3.0
azure-cosmos==4.5.1
opencensus-ext-azure==1.1.13
flask-cors==4.0.0
marshmallow==3.20.2
apispec==6.4.0
flasgger==0.9.7.1
